feat: append gif search result to webpage
Added an area for appending gifs in the HTML.

Created the appendGif method to add gifs to webpage.

In the searchGif method, changed the returned URL to be the one from
the images object, because the embedded URL was not working and was not
the correct one.  Also removed the console logs, because they were no
longer needed.

Changed handleSubmit method to be asynchronous and waited for
searchGif, because the URL returned from searchGif wasn't resolving to
a URL and was still a Promise by the time the appendGif method ran.